requirements.txt
beautifulsoup4==4.12.2
joblib==1.3.2
lxml==4.9.3
nltk==3.8.1
openai==1.3.7
pandas==1.5.3
pinecone-client==2.2.4
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
tdqm==0.0.1